.history/requirements_20251116194406.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0

# Media Processing
ffmpeg-python==0.2.0
# Pillow pin updated for Python 3.13 compatibility; original 10.1.0 lacked wheels for this interpreter.
Pillow==12.0.0
python-magic==0.4.27
pymediainfo==6.1.0

# Async & Concurrency
aiofiles==23.2.1
aiohttp==3.9.1

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# CLI
typer==0.9.0
rich==13.7.0
click==8.1.7

# Torrent (Phase 3)
# python-libtorrent==2.0.9  # Uncomment when implementing torrent features

# Compression
zstandard==0.22.0

# Search
whoosh==2.7.4

# Utilities
python-dateutil==2.8.2
humanize==4.9.0
tqdm==4.66.1
chardet==5.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14